+++ crates/smtp/src/inbound/data.rs
@@ -36,7 +36,7 @@ use mail_auth::{
 use mail_builder::headers::{date::Date, message_id::generate_message_id_header};
 use mail_parser::{MessageParser, parsers::fields::thread::thread_name};
 use registry::schema::structs::Rate;
-use sieve::runtime::Variable;
+use sieve::{SpamStatus, runtime::Variable};
 use smtp_proto::{
     MAIL_BY_RETURN, RCPT_NOTIFY_DELAY, RCPT_NOTIFY_FAILURE, RCPT_NOTIFY_NEVER, RCPT_NOTIFY_SUCCESS,
 };
@@ -416,6 +416,7 @@ impl<T: SessionStream> Session<T> {
 
         // Run SPAM filter
         let mut train_spam = None;
+        let mut spam_status = None;
         if self.server.core.spam.enabled
             && self
                 .server
@@ -441,6 +442,11 @@ impl<T: SessionStream> Session<T> {
                             is_spam,
                             thread_name(parsed_message.subject().unwrap_or_default()).to_string(),
                         )
+                    });
+                    spam_status = Some(if score.is_spam {
+                        SpamStatus::Spam
+                    } else {
+                        SpamStatus::Ham
                     });
 
                     // Add scores for local recipients
@@ -532,9 +538,13 @@ impl<T: SessionStream> Session<T> {
                     .map(|s| (s, name))
             })
         {
-            let params = self
+            let mut params = self
                 .build_script_parameters("data")
-                .with_auth_headers(&headers)
+                .with_auth_headers(&headers);
+            if let Some(spam_status) = spam_status {
+                params = params.with_spam_status(spam_status);
+            }
+            let params = params
                 .set_variable(
                     "arc.result",
                     arc_output